+++ src/codec_aom.c
@@ -693,24 +693,11 @@ static avifResult aomCodecEncodeImage(avifCodec * codec,
     // aom_codec.h says: aom_codec_version() == (major<<16 | minor<<8 | patch)
     static const int aomVersion_2_0_0 = (2 << 16);
     const int aomVersion = aom_codec_version();
-    if ((aomVersion < aomVersion_2_0_0) && (image->depth > 8)) {
-        // Due to a known issue with libaom v1.0.0-errata1-avif, 10bpc and
-        // 12bpc image encodes will call the wrong variant of
-        // aom_subtract_block when cpu-used is 7 or 8, and crash. Until we get
-        // a new tagged release from libaom with the fix and can verify we're
-        // running with that version of libaom, we must avoid using
-        // cpu-used=7/8 on any >8bpc image encodes.
-        //
-        // Context:
-        //   * https://github.com/AOMediaCodec/libavif/issues/49
-        //   * https://bugs.chromium.org/p/aomedia/issues/detail?id=2587
-        //
-        // Continued bug tracking here:
-        //   * https://github.com/AOMediaCodec/libavif/issues/56
-
-        if (aomCpuUsed > 6) {
-            aomCpuUsed = 6;
-        }
+    if (aomVersion <= aomVersion_2_0_0) {
+        // Issue with v1.0.0-errata1-avif: https://github.com/AOMediaCodec/libavif/issues/56
+        // Issue with v2.0.0: https://aomedia-review.googlesource.com/q/I26a39791f820b4d4e1d63ff7141f594c3c7181f5
+        avifDiagnosticsPrintf(codec->diag, "a libaom version strictly greater than 2.0.0 is required");
+        return AVIF_RESULT_UNKNOWN_ERROR;
     }
 
     avifBool useLibavifDefaultTuneMetric = AVIF_FALSE;        // If true, override libaom's default tune option.
@@ -872,19 +859,9 @@ static avifResult aomCodecEncodeImage(avifCodec * codec,
         }
 
         codec->internal->monochromeEnabled = AVIF_FALSE;
-        if (aomVersion > aomVersion_2_0_0) {
-            // There exists a bug in libaom's chroma_check() function where it will attempt to
-            // access nonexistent UV planes when encoding monochrome at faster libavif "speeds". It
-            // was fixed shortly after the 2.0.0 libaom release, and the fix exists in both the
-            // master and applejack branches. This ensures that the next version *after* 2.0.0 will
-            // have the fix, and we must avoid cfg->monochrome until then.
-            //
-            // Bugfix Change-Id: https://aomedia-review.googlesource.com/q/I26a39791f820b4d4e1d63ff7141f594c3c7181f5
-
-            if (alpha || (image->yuvFormat == AVIF_PIXEL_FORMAT_YUV400)) {
-                codec->internal->monochromeEnabled = AVIF_TRUE;
-                cfg->monochrome = 1;
-            }
+        if (alpha || (image->yuvFormat == AVIF_PIXEL_FORMAT_YUV400)) {
+            codec->internal->monochromeEnabled = AVIF_TRUE;
+            cfg->monochrome = 1;
         }
 
         if (!avifProcessAOMOptionsPreInit(codec, alpha, cfg)) {